pyproject.toml
[project]
name = "lichessko"
version = "0.1.0"
description = ""
authors = [
    {name = "Jaka Krajnc",email = "444.ous@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.4,<3.0.0)",
    "python-telegram-bot (>=22.5,<23.0)",
]

[tool.poetry]
packages = [{include = "lichessko", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
